修改帧率为浮点数

+++ FaceFeature/Handlers/DetectHandler.cs
@@ -101,7 +101,7 @@ public static class DetectHandler
         HttpContext context,
         DetectService detectService,
         ILogger<Program> logger,
-        int frameIntervalSeconds = 5,
+        double frameIntervalSeconds = 5,
         CancellationToken cancellationToken = default)
     {
         return HandleVideoAsync(context.Request, detectService, VideoCodec.H264, frameIntervalSeconds, logger, cancellationToken);
@@ -119,7 +119,7 @@ public static class DetectHandler
         HttpContext context,
         DetectService detectService,
         ILogger<Program> logger,
-        int frameIntervalSeconds = 5,
+        double frameIntervalSeconds = 5,
         CancellationToken cancellationToken = default)
     {
         return HandleVideoAsync(context.Request, detectService, VideoCodec.H265, frameIntervalSeconds, logger, cancellationToken);
@@ -132,7 +132,7 @@ public static class DetectHandler
         HttpRequest request,
         DetectService detectService,
         VideoCodec codec,
-        int frameIntervalSeconds,
+        double frameIntervalSeconds,
         ILogger logger,
         [EnumeratorCancellation] CancellationToken cancellationToken)
     {
